requirements.txt
pyCraft
openai
pyyaml
sqlite-utils
streamlit
watchdog 